requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.4.0=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.2.0=pypi_0
argon2-cffi=20.1.0=py36h1d69622_2
async-timeout=4.0.2=pypi_0
async_generator=1.10=py_0
asynctest=0.13.0=pypi_0
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=2.0.0=pyhd8ed1ab_0
blas=1.0=mkl
bleach=3.2.1=pyh9f0ad1d_0
blinker=1.4=py_1
boto3=1.21.32=pyhd8ed1ab_0
botocore=1.24.32=pyhd3eb1b0_0
bravado=11.0.3=pyhd8ed1ab_0
bravado-core=5.17.0=pyh9f0ad1d_0
brotlipy=0.7.0=py36he6145b8_1001
bzip2=1.0.8=h516909a_3
ca-certificates=2023.11.17=hbcca054_0
cachetools=4.2.4=pypi_0
cairo=1.14.12=h8948797_3
certifi=2021.5.30=py36h5fab9bb_0
cffi=1.14.3=py36h261ae71_2
chardet=3.0.4=py36h9880bd3_1008
charset-normalizer=3.0.1=pypi_0
click=8.0.1=py36h5fab9bb_0
cloudpickle=1.6.0=py_0
contextvars=2.4=pypi_0
cryptography=3.1.1=py36h5d711a5_1
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.10.0=py_2
dataclasses=0.8=pyh787bdff_2
dbus=1.13.6=he372182_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docstring-parser=0.15=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.2.9=he1b5a44_2
ffmpeg=4.0.2=ha0c5888_2
fontconfig=2.13.1=he4413a7_1000
freeglut=3.0.0=hf484d3e_1005
freetype=2.10.4=h7ca028e_0
frozenlist=1.2.0=pypi_0
fsspec=2022.1.0=pypi_0
future=0.18.2=py36h5fab9bb_3
fuzzywuzzy=0.18.0=pypi_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd8ed1ab_0
glib=2.66.1=h92f7085_0
gmp=6.1.2=hf484d3e_1000
gnutls=3.5.19=h2a4e5f8_1
google-auth=2.22.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
graphite2=1.3.14=h295c915_1
grpcio=1.48.2=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
h5py=2.8.0=py36h3010b51_1003
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
icu=58.2=hf484d3e_1000
idna=2.10=pyh9f0ad1d_0
idna-ssl=1.1.0=pypi_0
immutables=0.19=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py36he448a4c_1
ipython=7.16.1=py36he448a4c_2
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
jasper=2.0.14=hd8c5072_2
jedi=0.17.2=py36h5fab9bb_1
jinja2=2.11.2=pyh9f0ad1d_0
jmespath=0.10.0=pyh9f0ad1d_0
joblib=0.17.0=py_0
jpeg=9b=habf39ab_1
jsonargparse=4.24.1=pypi_0
jsonref=0.2=py_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py36h5fab9bb_2
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
keras=2.1.2=py36_0
kiwisolver=1.3.1=py36h51d7077_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_8
libblas=3.8.0=21_mkl
libcblas=3.8.0=21_mkl
libedit=3.1.20191231=he28a2e2_2
libffi=3.3=h58526e2_2
libgcc-ng=13.2.0=h807b86a_3
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_5
libglu=9.0.0=he1b5a44_1001
libiconv=1.16=h516909a_0
liblapack=3.8.0=21_mkl
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libuv=1.40.0=hd18ef5c_0
libvpx=1.7.0=h439df22_0
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=hb55368b_3
libzlib=1.2.11=h36c2ea0_1013
lightning-flash=0.7.5=pypi_0
llvm-openmp=14.0.4=he0ac6c6_0
lz4-c=1.9.2=he1b5a44_3
markdown=3.3.7=pypi_0
markupsafe=1.1.1=py36he6145b8_2
matplotlib=3.3.3=py36h5fab9bb_0
matplotlib-base=3.3.3=py36he12231b_0
mistune=0.8.4=pyh1a96a4e_1006
mkl=2020.2=ha770c72_256
mkl-service=2.3.0=py36h8c4c3a4_2
mkl_fft=1.2.0=py36h68bb277_1
mkl_random=1.1.1=py36h830a2c2_0
mock=4.0.3=py36h5fab9bb_1
monotonic=1.5=py_0
msgpack-python=1.0.2=py36hff7bd54_1
multidict=5.2.0=pypi_0
nb_conda=2.2.1=py36h5fab9bb_4
nb_conda_kernels=2.3.0=py36h9f0ad1d_3
nbclient=0.5.1=py_0
nbconvert=6.0.7=py36h5fab9bb_3
nbformat=5.0.8=py_0
ncurses=6.2=h58526e2_4
neptune-client=0.14.2=pyhd8ed1ab_0
nest-asyncio=1.4.1=py_0
nettle=3.3=0
networkx=2.5.1=pyhd8ed1ab_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
notebook=6.1.4=py36h9f0ad1d_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.0.0=py36_0
olefile=0.46=pyh9f0ad1d_1
opencv=3.4.2=py36h6fd60c2_1
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1w=hd590300_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.5=pypi_0
pandoc=2.11.4=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.7.0=pyh9f0ad1d_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=8.0.0=py36h9a89aac_0
pip=20.3.3=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.8=pyha770c72_0
protobuf=3.17.2=py36h295c915_0
psutil=5.8.0=py36h8f6f2f9_1
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.6.0=py_1001
py-opencv=3.4.2=py36hb342d67_1
py3nvml=0.2.5=py36h9f0ad1d_2
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pydeprecate=0.3.1=pypi_0
pygments=2.7.1=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyhd8ed1ab_1
pyqt=5.9.2=py36hcca6a23_4
pyrsistent=0.17.3=py36h1d69622_1
pysocks=1.7.1=py36h5fab9bb_3
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python_abi=3.6=2_cp36m
pytorch=1.9.0=py3.6_cuda11.1_cudnn8.0.5_0
pytorch-lightning=1.5.10=pypi_0
pytz=2020.1=pyh9f0ad1d_0
pyyaml=5.4.1=py36h27cfd23_1
pyzmq=19.0.2=py36h9947dbf_2
qt=5.9.7=h5867ecd_1
readline=8.0=he28a2e2_2
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.9=pypi_0
s3transfer=0.5.2=pyhd8ed1ab_0
scikit-learn=0.24.2=py36ha9443f7_0
scipy=1.5.2=py36h3a855aa_0
send2trash=1.5.0=py_0
setuptools=59.5.0=pypi_0
simplejson=3.17.3=py36h7f8727e_2
sip=4.19.8=py36hf484d3e_1000
six=1.15.0=pyh9f0ad1d_0
smmap=4.0.0=pyhd8ed1ab_1
sqlite=3.33.0=h4cf870e_1
swagger-spec-validator=2.7.3=pyh9f0ad1d_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=1.1.0=py36_0
terminado=0.9.1=py36h5fab9bb_1
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h21135ba_1
torchaudio=0.9.0=py36
torchmetrics=0.6.2=pyhd8ed1ab_0
torchvision=0.2.1=py36_0
tornado=6.0.4=py36h8c4c3a4_2
tqdm=4.55.1=pyhd8ed1ab_0
traitlets=4.3.3=pyhd8ed1ab_2
typeshed-client=2.1.0=pypi_0
typing-extensions=3.10.0.0=hd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
urllib3=1.25.11=py_0
warmup-scheduler=0.3=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=0.58.0=py36h06a4308_4
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3deb0d_0
widgetsnbextension=3.5.1=py36h5fab9bb_4
x264=1!152.20180806=h14c3975_0
xmltodict=0.12.0=py_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h14c3975_1002
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libx11=1.6.12=h36c2ea0_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarl=1.7.2=pypi_0
zeromq=4.3.3=h58526e2_3
zipp=3.3.1=py_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.4.5=h6597ccf_2
